+++ rapidf-components/rapidf-auapi/src/main/java/com/lazycece/rapidf/auapi/extension/ApiParams.java
@@ -1,0 +1,50 @@
+/*
+ *    Copyright 2023 lazycece<lazycece@gmail.com>
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package com.lazycece.rapidf.auapi.extension;
+
+import org.apache.commons.lang3.StringUtils;
+
+/**
+ * @author lazycece
+ */
+public class ApiParams extends com.lazycece.au.api.params.ApiParams {
+    private String deviceId;
+    private String version;
+
+    public String getDeviceId() {
+        return deviceId;
+    }
+
+    public void setDeviceId(String deviceId) {
+        this.deviceId = deviceId;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
+    }
+
+    @Override
+    public boolean validate() {
+        return super.validate()
+                && StringUtils.isNotBlank(deviceId)
+                && StringUtils.isNotBlank(version);
+    }
+}
